add `*_entry.rs` files
These will have the same lifetime as the `Manifest` struct, and are
going to be available in the public api.

I feel like providing vectors for each of these will severely improve
the api, because the goal is not to provide just the files a user may
download, but instead provide all of the info the file contains.

I'm making custom structs, instead of just using the ones in the
generated file, for three reasons:
1. Generated structures are only valid for the lifetime of decompressed
data. Decompressed data would need to be stored in one of the structs
becuse of that.
2. Making custom structures allows me to verify all of the data during
parsing, and not when needed.
3. Generated structures (from my understanding) read the data when
requested. Because of this, multiple reads over the values are
theoretically slower.

The structure for `file_entries` is comming in the following commit.