add rman-schema as a submodule
We will be generating flatbuffer schema on-demand using git submodules
and rust build scripts.

This will make updating this crate easier if the schema changes and
allows us to not have a generated rust in the project directory (this
wasn't particulary great because of all the clippy and other rules that
had to be disabled for that file specifically).

Build script is comming in the next commit.