compile flatbuffer schema using build scripts
Didn't go quite the way I expected it to, but it's still better.
Originally I had hoped that this way adding various clippy allow rules
for the generated schema would not be required, but that is not the case.
But now those are no longer in the generated file which was the goal.

Four tests are now broken, because we no longer have any way to generate
test manifest files.