convert `ManifestFile` unit tests into integration tests
This way, we can have setup functions which generate manifest files used
for testing. We include the parsed flatbuffer schema the same way the
library does, because it's visibilty is private.

We also don't clean up these files intentionally, because they will
hopefully be able to be used by doc tests (this is with the assumption
that doc tests are always ran after integration tests).